@@ -1,7 +1,8 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
 import style from './NavBar.module.css'
-export default function NavBar() {
+export default function NavBar(props) {
+  console.log(props.userData)
   return (
     <div >
       <nav className={`navbar navbar-expand-lg navbar-dark ${style.bgColor}`}>
@@ -12,41 +13,53 @@ export default function NavBar() {
           </button>
           <div className="collapse navbar-collapse" id="navbarSupportedContent">
             <ul className="navbar-nav me-auto mb-2 mb-lg-0">
-              <li className="nav-item">
-                <Link className="nav-link active" aria-current="page" to='/home' >Home</Link>
-              </li>
-              <li className="nav-item">
-                <Link className="nav-link active" aria-current="page" to="/movies">Movies</Link>
-              </li>
-              <li className="nav-item">
-                <Link className="nav-link active" aria-current="page" to="/tvshows">TvShows</Link>
-              </li>
-              <li className="nav-item">
-                <Link className="nav-link active" aria-current="page" to="people">People</Link>
-              </li>
-              <li className="nav-item">
-                <Link className="nav-link active" aria-current="page" to="network">Network</Link>
-              </li>
+              {
+                props.userData ? '' 
+                :<>
+                  <li className="nav-item">
+                    <Link className="nav-link active" aria-current="page" to='/home' >Home</Link>
+                  </li>
+                  <li className="nav-item">
+                    <Link className="nav-link active" aria-current="page" to="/movies">Movies</Link>
+                  </li>
+                  <li className="nav-item">
+                    <Link className="nav-link active" aria-current="page" to="/tvshows">TvShows</Link>
+                  </li>
+                  <li className="nav-item">
+                    <Link className="nav-link active" aria-current="page" to="people">People</Link>
+                  </li>
+                  <li className="nav-item">
+                    <Link className="nav-link active" aria-current="page" to="network">Network</Link>
+                  </li>
+                </> 
+              }
 
             </ul>
             <ul className="navbar-nav ms-auto mb-2 mb-lg-0">
               <li>
-                <a href="">
+                <a href="www.facebook.com">
                   <i class="fa-brands fa-facebook-square"></i>
                 </a>
               </li>
               <li>
-                <a href=""><i class="fa-brands fa-instagram-square"></i></a>
+                <a href="www.instagram.com"><i class="fa-brands fa-instagram-square"></i></a>
               </li>
               <li>
-                <a href=""><i class="fa-brands fa-twitter-square"></i></a>
+                <a href="www.twitter.com"><i class="fa-brands fa-twitter-square"></i></a>
               </li>
-              <li className='nav-item'>
-                <Link className="nav-link active" to={"register"} >Register</Link>
-              </li>
-              <li className='nav-item'>
-                <Link className="nav-link active" to={"signin"} >SignIn</Link>
-              </li>
+              {
+                props.userData ? 
+                   <a onClick={props.signOut} className="nav-link active">LogOut</a>
+                  : <>
+                  <li className='nav-item'>
+                    <Link className="nav-link active" to={"register"} >Register</Link>
+                  </li>
+                  <li className='nav-item'>
+                    <Link className="nav-link active" to={"signin"} >SignIn</Link>
+                  </li>
+                  <li className='nav-item'></li>
+                </>
+              }
             </ul>
 
           </div>
